made code review changes
