+++ src/main/java/org/usfirst/frc/team5968/robot/FieldPosition.java
@@ -5,5 +5,4 @@ public enum FieldPosition {
     CENTER,
     RIGHT,
     INVALID
-
 }
