+++ src/main/java/org/usfirst/frc/team5968/robot/HABLineAuto.java
@@ -17,7 +17,6 @@ public class HABLineAuto implements IRobotMode {
 
     @Override
     public void init() {
-        robotSpeed = LOW;
         driveStraight();
     }
