--- src/main/java/org/usfirst/frc/team5968/robot/HardCodedDashboard.java
@@ -1,20 +1,0 @@
-package org.usfirst.frc.team5968.robot;
-/*
-public class HardCodedDashboard implements IFieldInformation {
-
-    @Override
-    public FieldPosition getStartingPosition() {
-        return startingPosition;
-
-    }
-
-    @Override
-    public boolean getIsDataValid() {
-
-    }
-
-    @Override
-    public void refresh() {
-
-    } */
-
