+++ src/main/java/org/usfirst/frc/team5968/robot/Robot.java
@@ -5,10 +5,6 @@ import edu.wpi.first.hal.HAL;
 import edu.wpi.first.wpilibj.livewindow.LiveWindow;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
-//import com.ctre.phoenix.motorcontrol.ControlMode;
-//import com.ctre.phoenix.motorcontrol.can.TalonSRX;
-//import edu.wpi.first.wpilibj.Joystick;
-
 public class Robot extends RobotBase {
 
     private IRobotMode disabledMode;
@@ -21,98 +17,9 @@ public class Robot extends RobotBase {
     private ICargoGuide cargoGuide;
     private IGyroscopeSensor gyroscope;
 
-    /*
-    private class MotorTest implements IRobotMode
-    {
-        private Boolean lastButton0 = false;
-        private Boolean lastButton1 = false;
-        private TalonSRX[] motors;
-        private int currentMotor = 0;
-        private Joystick joystick;
-
-        private Boolean lastDoubleButton = false;
-        private Boolean doubleMode = false;
-
-        public MotorTest() {
-            int[] canPorts = {
-                10, 3, 6, 4, 5, 7
-            };
-
-            motors = new TalonSRX[canPorts.length];
-
-            for (int i = 0; i < canPorts.length; i++) {
-                motors[i] = new TalonSRX(canPorts[i]);
-            }
-
-            joystick = new Joystick(0);
-        }
-
-        @Override
-        public void init()
-        {
-            currentMotor = 0;
-            lastButton0 = false;
-            lastButton1 = false;
-            doubleMode = false;
-            lastDoubleButton = false;
-        }
-
-        @Override
-        public void periodic()
-        {
-            Boolean button0 = joystick.getRawButton(11);
-            Boolean button1 = joystick.getRawButton(12);
-            Boolean doubleButton = joystick.getRawButton(7);
-
-            if (button0 && !lastButton0) {
-                currentMotor--;
-
-                if (currentMotor < 0) {
-                    currentMotor = motors.length - 1;
-                }
-            } else if (button1 && !lastButton1) {
-                currentMotor++;
-
-                if (currentMotor >= motors.length) {
-                    currentMotor = 0;
-                }
-            }
-
-            if (doubleButton && !lastDoubleButton) {
-                doubleMode = !doubleMode;
-            }
-
-            lastButton0 = button0;
-            lastButton1 = button1;
-            lastDoubleButton = doubleButton;
-            double speed = joystick.getY();
-            motors[currentMotor].set(ControlMode.PercentOutput, speed);
-            int nextMotorIndex = -1;
-
-            if (doubleMode) {
-                nextMotorIndex = currentMotor + 1;
-
-                if (nextMotorIndex >= motors.length) {
-                    nextMotorIndex = 0;
-                }
-
-                motors[nextMotorIndex].set(ControlMode.PercentOutput, speed);
-            }
-
-            for (int i = 0; i < motors.length; i++) {
-                if (i == currentMotor || i == nextMotorIndex) {
-                    continue;
-                }
-
-                motors[i].set(ControlMode.PercentOutput, 0.0);
-            }
-        }
-    } */
-
     public Robot() {
         gyroscope = new NavXMXP();
         drive = new Drive(gyroscope);
-        //drive = new NullDrive();
         hook = new Hook();
         launcher = new Launcher();
         cargoGuide = new CargoGuide();
@@ -120,7 +27,6 @@ public class Robot extends RobotBase {
         disabledMode = new DisabledMode(hook, launcher);
         autonomousMode = new AutonomousMode(drive, hook);
         teleoperatedMode = new TeleoperatedMode(drive, hook, launcher, cargoGuide);
-        //teleoperatedMode = new MotorTest();
     }
 
     @Override
@@ -148,15 +54,16 @@ public class Robot extends RobotBase {
 
 
     private void doPeripheralReinitialization() {
-        drive.init();
-        launcher.init();
-        hook.init();
         cargoGuide.init();
+        drive.init();
+        hook.init();
+        launcher.init();
     }
 
     private void doPeripheralPeriodicProcessing() {
         drive.periodic();
         launcher.periodic();
+
         Debug.periodic();
     }
